requirements.txt
asgiref==3.6.0
Django==3.2.19
django-cors-headers==4.0.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==2.3.0
future==0.18.3
importlib-metadata==6.6.0
Markdown==3.4.3
mccabe==0.7.0
packaging==23.1
pep8==1.7.1
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.7.0
pytz==2023.3
serializers==0.2.4
sqlparse==0.4.4
typing_extensions==4.5.0
uuid==1.30
zipp==3.15.0
